+++ storybook-static/assets/preview-00d99885.js
@@ -1,2 +1,0 @@
-import{_ as e}from"./iframe-d2951414.js";import"../sb-preview/runtime.js";var a={docs:{renderer:async()=>{let{DocsRenderer:r}=await e(()=>import("./DocsRenderer-3PUGWF3O-8a85424e.js"),["./DocsRenderer-3PUGWF3O-8a85424e.js","./chunk-S4VUQJ4A-d919f0b3.js","./iframe-d2951414.js","./index-d8ade8ea.js","./_commonjsHelpers-042e6b4d.js","./index-d475d2ea.js","./index-d37d4223.js","./index-6ba087f2.js","./floating-ui.dom-c3a31ea6.js","./index-356e4a49.js"],import.meta.url);return new r}}};export{a as parameters};
-//# sourceMappingURL=preview-00d99885.js.map